+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_format.py
@@ -5,125 +5,137 @@
 
 from collections import OrderedDict
 
-_property_map = {
-    'name': 'NAME',
-    'resourceGroup': 'RESOURCE GROUP',
-    'location': 'LOCATION',
-    'loginServer': 'LOGIN SERVER',
-    'creationDate': 'CREATION DATE',
-    'adminUserEnabled': 'ADMIN ENABLED',
-    'status': 'STATUS',
-    'scope': 'SCOPE',
-    'actions': 'ACTIONS',
-    'serviceUri': 'SERVICE URI',
-    'customHeaders': 'HEADERS',
-    'limit': 'LIMIT',
-    'currentValue': 'CURRENT VALUE',
-    'unit': 'UNIT'
-}
 
-_order_map = {
-    'NAME': 1,
-    'RESOURCE GROUP': 2,
-    'LOCATION': 3,
-    'SKU': 4,
-    'LOGIN SERVER': 11,
-    'CREATION DATE': 12,
-    'ADMIN ENABLED': 13,
-    'USERNAME': 31,
-    'PASSWORD': 32,
-    'PASSWORD2': 33,
-    'STATUS': 41,
-    'SCOPE': 42,
-    'ACTIONS': 43,
-    'SERVICE URI': 44,
-    'HEADERS': 45,
-    'LIMIT': 51,
-    'CURRENT VALUE': 52,
-    'UNIT': 53,
-    'ID': 61,
-    'ACTION': 62,
-    'IMAGE': 63,
-    'RESPONSE STATUS': 64,
-    'TIMESTAMP': 65
-}
+def registry_output_format(result):
+    return _output_format(result, _registry_format_group)
 
 
-def output_format(result):
-    """Returns the list of container registries each of which is an ordered dictionary.
-    :param list/dict result: The (list of) container registry object(s)
-    """
+def usage_output_format(result):
+    return _output_format(result, _usage_format_group)
+
+
+def credential_output_format(result):
+    return _output_format(result, _credential_format_group)
+
+
+def webhook_output_format(result):
+    return _output_format(result, _webhook_format_group)
+
+
+def webhook_get_config_output_format(result):
+    return _output_format(result, _webhook_get_config_format_group)
+
+
+def webhook_list_events_output_format(result):
+    return _output_format(result, _webhook_list_events_format_group)
+
+
+def webhook_ping_output_format(result):
+    return _output_format(result, _webhook_ping_format_group)
+
+
+def replication_output_format(result):
+    return _output_format(result, _replication_format_group)
+
+
+def _output_format(result, format_group):
     if 'value' in result and isinstance(result['value'], list):
         result = result['value']
     obj_list = result if isinstance(result, list) else [result]
-    return [_format_group(item) for item in obj_list]
+    return [format_group(item) for item in obj_list]
 
 
-def _format_group(item):
-    """Returns an ordered dictionary of the container registry.
-    :param dict item: The container registry object
+def _registry_format_group(item):
+    return OrderedDict([
+        ('NAME', _get_value(item, 'name')),
+        ('RESOURCE GROUP', _get_value(item, 'resourceGroup')),
+        ('LOCATION', _get_value(item, 'location')),
+        ('SKU', _get_value(item, 'sku', 'name')),
+        ('LOGIN SERVER', _get_value(item, 'loginServer')),
+        ('CREATION DATE', _format_datetime(_get_value(item, 'creationDate'))),
+        ('ADMIN ENABLED', _get_value(item, 'adminUserEnabled'))
+    ])
+
+
+def _usage_format_group(item):
+    return OrderedDict([
+        ('NAME', _get_value(item, 'name')),
+        ('LIMIT', _get_value(item, 'limit')),
+        ('CURRENT VALUE', _get_value(item, 'currentValue')),
+        ('UNIT', _get_value(item, 'unit'))
+    ])
+
+
+def _credential_format_group(item):
+    return OrderedDict([
+        ('USERNAME', _get_value(item, 'username')),
+        ('PASSWORD', _get_value(item, 'passwords', 0, 'value')),
+        ('PASSWORD2', _get_value(item, 'passwords', 1, 'value'))
+    ])
+
+
+def _webhook_format_group(item):
+    return OrderedDict([
+        ('NAME', _get_value(item, 'name')),
+        ('LOCATION', _get_value(item, 'location')),
+        ('ACTIONS', _get_value(item, 'actions')),
+        ('SCOPE', _get_value(item, 'scope')),
+        ('STATUS', _get_value(item, 'status'))
+    ])
+
+
+def _webhook_get_config_format_group(item):
+    return OrderedDict([
+        ('SERVICE URI', _get_value(item, 'serviceUri')),
+        ('HEADERS', _get_value(item, 'customHeaders'))
+    ])
+
+
+def _webhook_list_events_format_group(item):
+    repository = _get_value(item, 'eventRequestMessage', 'content', 'target', 'repository').strip()
+    tag = _get_value(item, 'eventRequestMessage', 'content', 'target', 'tag').strip()
+    status = _get_value(item, 'eventResponseMessage', 'statusCode').strip()
+    reason = _get_value(item, 'eventResponseMessage', 'reasonPhrase').strip()
+
+    return OrderedDict([
+        ('ID', _get_value(item, 'id')),
+        ('ACTION', _get_value(item, 'eventRequestMessage', 'content', 'action')),
+        ('IMAGE', '{}:{}'.format(repository, tag) if repository and tag else repository or ' '),
+        ('HTTP STATUS', '{} {}'.format(status, reason) if status and reason else status or reason or ' '),
+        ('TIMESTAMP', _format_datetime(_get_value(item, 'eventRequestMessage', 'content', 'timestamp')))
+    ])
+
+
+def _webhook_ping_format_group(item):
+    return OrderedDict([
+        ('ID', _get_value(item, 'id'))
+    ])
+
+
+def _replication_format_group(item):
+    return OrderedDict([
+        ('NAME', _get_value(item, 'name')),
+        ('LOCATION', _get_value(item, 'location')),
+        ('PROVISIONING STATE', _get_value(item, 'provisioningState')),
+        ('STATUS', _get_value(item, 'status', 'displayStatus'))
+    ])
+
+
+def _format_datetime(date_string):
+    from dateutil.parser import parse
+    try:
+        return parse(date_string).strftime("%Y-%m-%dT%H:%M:%SZ")
+    except ValueError:
+        return date_string or ' '
+
+
+def _get_value(item, *args):
+    """Recursively get a nested value from a dict.
+    :param dict item: The dict object
     """
-    table_info = {_property_map[key]: str(item[key]) for key in item if key in _property_map}
-
     try:
-        table_info['SKU'] = item['sku']['name']
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        table_info['USERNAME'] = item['username']
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        table_info['PASSWORD'] = item['passwords'][0]['value']
+        for arg in args:
+            item = item[arg]
+        return str(item) if item else ' '
     except (KeyError, TypeError, IndexError):
-        pass
-
-    try:
-        table_info['PASSWORD2'] = item['passwords'][1]['value']
-    except (KeyError, TypeError, IndexError):
-        pass
-
-    try:
-        # Only show ID if it is not an ARM resource ID
-        table_info['ID'] = item['id'] if '/subscriptions/' not in item['id'].lower() else None
-    except (KeyError, TypeError):
-        pass
-
-    # Parse webhook list-events
-    try:
-        table_info['ACTION'] = item['eventRequestMessage']['content']['action']
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        table_info['IMAGE'] = item['eventRequestMessage']['content']['target']['repository']
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        tag = item['eventRequestMessage']['content']['target']['tag']
-        if table_info['IMAGE'] and tag:
-            table_info['IMAGE'] = '{}:{}'.format(table_info['IMAGE'], tag)
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        table_info['TIMESTAMP'] = item['eventRequestMessage']['content']['timestamp']
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        table_info['RESPONSE STATUS'] = item['eventResponseMessage']['statusCode']
-    except (KeyError, TypeError):
-        pass
-
-    try:
-        status_code = item['eventResponseMessage']['statusCode']
-        reason_phrase = item['eventResponseMessage']['reasonPhrase']
-        table_info['RESPONSE STATUS'] = '{} {}'.format(status_code, reason_phrase)
-    except (KeyError, TypeError):
-        pass
-
-    return OrderedDict(sorted(table_info.items(), key=lambda t: _order_map[t[0]]))
+        return ' '