+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/commands.py
@@ -6,7 +6,16 @@
 from azure.cli.core.commands import CliCommandType
 from azure.cli.core.util import empty_on_404
 
-from ._format import output_format
+from ._format import (
+    registry_output_format,
+    usage_output_format,
+    credential_output_format,
+    webhook_output_format,
+    webhook_get_config_output_format,
+    webhook_list_events_output_format,
+    webhook_ping_output_format,
+    replication_output_format
+)
 from ._client_factory import cf_acr_registries, cf_acr_replications, cf_acr_webhooks
 
 
@@ -14,13 +23,13 @@ def load_command_table(self, _):
 
     acr_custom_util = CliCommandType(
         operations_tmpl='azure.cli.command_modules.acr.custom#{}',
-        table_transformer=output_format,
+        table_transformer=registry_output_format,
         client_factory=cf_acr_registries
     )
 
     acr_cred_util = CliCommandType(
         operations_tmpl='azure.cli.command_modules.acr.credential#{}',
-        table_transformer=output_format,
+        table_transformer=credential_output_format,
         client_factory=cf_acr_registries
     )
 
@@ -30,13 +39,13 @@ def load_command_table(self, _):
 
     acr_webhook_util = CliCommandType(
         operations_tmpl='azure.cli.command_modules.acr.webhook#{}',
-        table_transformer=output_format,
+        table_transformer=webhook_output_format,
         client_factory=cf_acr_webhooks
     )
 
     acr_replication_util = CliCommandType(
         operations_tmpl='azure.cli.command_modules.acr.replication#{}',
-        table_transformer=output_format,
+        table_transformer=replication_output_format,
         client_factory=cf_acr_replications
     )
 
@@ -46,14 +55,15 @@ def load_command_table(self, _):
         g.command('create', 'acr_create')
         g.command('delete', 'acr_delete')
         g.command('show', 'acr_show', exception_handler=empty_on_404)
-        g.command('login', 'acr_login')
-        g.command('show-usage', 'acr_show_usage')
+        g.command('login', 'acr_login', table_transformer=None)
+        g.command('show-usage', 'acr_show_usage', table_transformer=usage_output_format)
         g.generic_update_command('update',
                                  getter_name='acr_update_get',
                                  setter_name='acr_update_set',
                                  custom_func_name='acr_update_custom',
                                  custom_func_type=acr_custom_util,
-                                 client_factory=cf_acr_registries)
+                                 client_factory=cf_acr_registries,
+                                 table_transformer=registry_output_format)
 
     with self.command_group('acr credential', acr_cred_util) as g:
         g.command('show', 'acr_credential_show', exception_handler=empty_on_404)
@@ -71,15 +81,16 @@ def load_command_table(self, _):
         g.command('create', 'acr_webhook_create')
         g.command('delete', 'acr_webhook_delete')
         g.command('show', 'acr_webhook_show')
-        g.command('get-config', 'acr_webhook_get_config')
-        g.command('list-events', 'acr_webhook_list_events')
-        g.command('ping', 'acr_webhook_ping')
+        g.command('get-config', 'acr_webhook_get_config', table_transformer=webhook_get_config_output_format)
+        g.command('list-events', 'acr_webhook_list_events', table_transformer=webhook_list_events_output_format)
+        g.command('ping', 'acr_webhook_ping', table_transformer=webhook_ping_output_format)
         g.generic_update_command('update',
                                  getter_name='acr_webhook_update_get',
                                  setter_name='acr_webhook_update_set',
                                  custom_func_name='acr_webhook_update_custom',
                                  custom_func_type=acr_webhook_util,
-                                 client_factory=cf_acr_webhooks)
+                                 client_factory=cf_acr_webhooks,
+                                 table_transformer=webhook_output_format)
 
     with self.command_group('acr replication', acr_replication_util) as g:
         g.command('list', 'acr_replication_list')
@@ -91,4 +102,5 @@ def load_command_table(self, _):
                                  setter_name='acr_replication_update_set',
                                  custom_func_name='acr_replication_update_custom',
                                  custom_func_type=acr_replication_util,
-                                 client_factory=cf_acr_replications)
+                                 client_factory=cf_acr_replications,
+                                 table_transformer=replication_output_format)