+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.0.24
 ++++++
 * Improve repository commands error messages.
+* Update table columns and ordering.
 
 2.0.23
 ++++++